no cambios, arreglando git timeline

+++ functions/msfm3d_MOD.h
@@ -1,9 +1,9 @@
 #ifndef MSFM2D_MOD_H
 #define MSFM2D_MOD_H
 
-    float* main_msfm3D(float* F, float* source_points, float* T, int* size_map, int* size_target,
-                        float* start_points);
-    void compute_gradient_3d_discrete(float* input_matrix, float* gradient_matrix, int* size_map);
-    float* velocities_map3D(float* matriz, int* size_map, int distance_threshold);
+    double* main_msfm3D(double* F, double* source_points, double* T, int* size_map, int* size_target,
+                        double* start_points);
+    void compute_gradient_3d_discrete(double* input_matrix, double* gradient_matrix, int* size_map);
+    double* velocities_map3D(double* matriz, int* size_map, int distance_threshold);
 
 #endif
